refactor(coins): move walkthrough classes to dedicated package
